requirements.txt
discord.py
gtts
Pillow
yt_dlp
openai
ffmpy